chore(iam): Add schema/types (#395)
