+++ internal/provider/iam/role_schema.go
@@ -16,21 +16,6 @@ import (
 	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
 )
 
-type roleResourceModel struct {
-	ID          types.String `tfsdk:"id"`
-	Name        types.String `tfsdk:"name"`
-	Description types.String `tfsdk:"description"`
-	Rights      types.Set    `tfsdk:"rights"`
-}
-
-type roleDataSourceModel struct {
-	ID          types.String `tfsdk:"id"`
-	Name        types.String `tfsdk:"name"`
-	Description types.String `tfsdk:"description"`
-	ReadOnly    types.Bool   `tfsdk:"read_only"`
-	Rights      types.Set    `tfsdk:"rights"`
-}
-
 /*
 roleSchema
